pyproject.toml
[project]
name = "lm-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths  = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:torch.*",
]
